New Stuff/Fixes
More details in the discord chat

+++ GameDesignProtoype/Assets/Scripts/killZone.cs
@@ -8,6 +8,8 @@ public class killZone : MonoBehaviour
     public GameObject minionDead1;
     public GameObject minionDead2;
     public GameObject minionDead3;
+    private bool tempactive = true;
+    public bool permanent = false;
 
     void Start()
     {
@@ -18,27 +20,34 @@ public class killZone : MonoBehaviour
 
     private void OnTriggerEnter2D(Collider2D other)
     {
-        if (other.gameObject.CompareTag("Player"))
+        if (tempactive || permanent) //If the deathplane hasn't been deactivated or is always active
         {
-            GameObject.Find("Player").SendMessage("gotoCP");
+            if (other.gameObject.CompareTag("Player"))
+            {
+                GameObject.Find("Player").SendMessage("gotoCP");
+            }
+
+            if (other.gameObject.CompareTag("Minion1"))
+            {
+                GameObject.FindWithTag("Minion1").SendMessage("DeactivateMinion");
+                minionDead1.SendMessage("changeColorDead");
+                tempactive = false; //This and other duplicate lines deactivate the death plane, but only when the minions touch it.
+                
+            }
+
+            if (other.gameObject.CompareTag("Minion2"))
+            {
+                GameObject.FindWithTag("Minion2").SendMessage("DeactivateMinion");
+                minionDead2.SendMessage("changeColorDead");
+                tempactive = false;
+            }
+
+            if (other.gameObject.CompareTag("Minion3"))
+            {
+                GameObject.FindWithTag("Minion3").SendMessage("DeactivateMinion");
+                minionDead3.SendMessage("changeColorDead");
+                tempactive = false;
+            } //These three perma-kill each minion when they touch it. Needed 1 for each minion because I don't know how to format it more efficiently
         }
-        
-        if (other.gameObject.CompareTag("Minion1"))
-        {
-            GameObject.FindWithTag("Minion1").SendMessage("DeactivateMinion");
-            minionDead1.SendMessage("changeColorDead");
-        }
-        
-        if (other.gameObject.CompareTag("Minion2"))
-        {
-            GameObject.FindWithTag("Minion2").SendMessage("DeactivateMinion");
-            minionDead2.SendMessage("changeColorDead");
-        }
-        
-        if (other.gameObject.CompareTag("Minion3"))
-        {
-            GameObject.FindWithTag("Minion3").SendMessage("DeactivateMinion");
-            minionDead3.SendMessage("changeColorDead");
-        } //These three perma-kill each minion when they touch it. Needed 1 for each minion because I don't know how to format it more efficiently
     }
 }
